Create new migrations including config table
Rearrange order that tables are created in the revision, and
delete older revisions.
