--- alembic/versions/0686c8b1ed0b_add_config_table.py
@@ -1,37 +1,0 @@
-"""add config table
-
-Revision ID: 0686c8b1ed0b
-Revises: b692dfd0b017
-Create Date: 2021-06-14 22:03:13.976750
-
-"""
-from alembic import op
-import sqlalchemy as sa
-import sqlalchemy_utils
-
-
-# revision identifiers, used by Alembic.
-revision = '0686c8b1ed0b'
-down_revision = 'b692dfd0b017'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('config',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('product', sa.String(), nullable=True),
-    sa.Column('features', sqlalchemy_utils.types.scalar_list.ScalarListType(), nullable=True),
-    sa.Column('license_servers', sqlalchemy_utils.types.scalar_list.ScalarListType(), nullable=True),
-    sa.Column('license_server_type', sa.String(), nullable=True),
-    sa.Column('grace_time', sa.Integer(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('config')
-    # ### end Alembic commands ###
