+++ alembic/versions/5579536ad0f6_add_foreign_key_to_config_table_in_the_.py
@@ -1,0 +1,63 @@
+"""add foreign key to config table in the booking table
+
+Revision ID: 5579536ad0f6
+Revises: 
+Create Date: 2021-07-07 22:10:03.229052
+
+"""
+from alembic import op
+import sqlalchemy as sa
+import sqlalchemy_utils
+
+
+# revision identifiers, used by Alembic.
+revision = '5579536ad0f6'
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('config',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('product', sa.String(), nullable=True),
+    sa.Column('features', sqlalchemy_utils.types.scalar_list.ScalarListType(), nullable=True),
+    sa.Column('license_servers', sqlalchemy_utils.types.scalar_list.ScalarListType(), nullable=True),
+    sa.Column('license_server_type', sa.String(), nullable=True),
+    sa.Column('grace_time', sa.Integer(), nullable=True),
+    sa.PrimaryKeyConstraint('id'),
+    sqlite_autoincrement=True
+    )
+    op.create_table('license',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('product_feature', sa.String(), nullable=False),
+    sa.Column('used', sa.Integer(), nullable=True),
+    sa.Column('total', sa.Integer(), nullable=True),
+    sa.CheckConstraint('used<=total'),
+    sa.PrimaryKeyConstraint('id', 'product_feature'),
+    sqlite_autoincrement=True
+    )
+    op.create_table('booking',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('job_id', sa.String(), nullable=True),
+    sa.Column('product_feature', sa.String(), nullable=True),
+    sa.Column('booked', sa.Integer(), nullable=True),
+    sa.Column('lead_host', sa.String(), nullable=True),
+    sa.Column('user_name', sa.String(), nullable=True),
+    sa.Column('cluster_name', sa.String(), nullable=True),
+    sa.Column('created_at', sa.DateTime(), nullable=True),
+    sa.Column('config_id', sa.Integer(), nullable=False),
+    sa.ForeignKeyConstraint(['config_id'], ['config.id'], ),
+    sa.PrimaryKeyConstraint('id'),
+    sqlite_autoincrement=True
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('booking')
+    op.drop_table('license')
+    op.drop_table('config')
+    # ### end Alembic commands ###
