--- alembic/versions/ee200fc08826_add_foreign_key_to_config_table_in_the_.py
@@ -1,54 +1,0 @@
-"""add foreign key to config table in the booking table
-
-Revision ID: ee200fc08826
-Revises: 0686c8b1ed0b
-Create Date: 2021-07-07 21:00:28.643614
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'ee200fc08826'
-down_revision = '0686c8b1ed0b'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('booking', sa.Column('id', sa.Integer(), nullable=False))
-    op.add_column('booking', sa.Column('lead_host', sa.String(), nullable=True))
-    op.add_column('booking', sa.Column('user_name', sa.String(), nullable=True))
-    op.add_column('booking', sa.Column('cluster_name', sa.String(), nullable=True))
-    op.add_column('booking', sa.Column('created_at', sa.DateTime(), nullable=True))
-    op.add_column('booking', sa.Column('config_id', sa.Integer(), nullable=True))
-    op.alter_column('booking', 'job_id',
-               existing_type=sa.VARCHAR(),
-               nullable=True)
-    op.alter_column('booking', 'product_feature',
-               existing_type=sa.VARCHAR(),
-               nullable=True)
-    op.create_foreign_key(None, 'booking', 'config', ['config_id'], ['id'])
-    op.add_column('license', sa.Column('id', sa.Integer(), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('license', 'id')
-    op.drop_constraint(None, 'booking', type_='foreignkey')
-    op.alter_column('booking', 'product_feature',
-               existing_type=sa.VARCHAR(),
-               nullable=False)
-    op.alter_column('booking', 'job_id',
-               existing_type=sa.VARCHAR(),
-               nullable=False)
-    op.drop_column('booking', 'config_id')
-    op.drop_column('booking', 'created_at')
-    op.drop_column('booking', 'cluster_name')
-    op.drop_column('booking', 'user_name')
-    op.drop_column('booking', 'lead_host')
-    op.drop_column('booking', 'id')
-    # ### end Alembic commands ###
